Initial commit
This script contains a class for generating the WikiProject category
tree. Initially this will be used for auditing the WikiProject category
system (“WikiProjects by area” and descendants) but will ultimately be
used to produce a tiered category list for use by directory.py.

In this initial version, it will only go two levels deep. It will need
to go deeper than that.