New script
The assessment predictor tool predicts article quality and priority for
a given WikiProject. The article quality part is powered by the ORES
artificial intelligence service; this is not yet implemented in my
script. What *is* implemented is the priority predictor. Predictions
are done per WikiProject and are based on a combination of page views
and inbound link count (with page views being given 75% weight to
inbound links’ 25%).